+++ src/utils/toFunction.ts
@@ -1,0 +1,2 @@
+export type ToFunction_F = <T = any>(value: T) => () => T;
+export const toFunction: ToFunction_F = value => () => value;