+++ CHANGELOG.md
@@ -8,7 +8,25 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <br/>
 
-## [0.1.4-beta] - 2024-03-14 _00:00_
+## [0.1.10-beta] - 2024-03-15 _17:00_
+
+- Fix library import @bemedev/decompose
+
+<br/>
+
+## [0.1.9-beta] - 2024-03-15 _15:10_
+
+- Simplify types for interpret
+
+<br/>
+
+## [0.1.5-beta] - 2024-03-15 _03:08_
+
+- Fix subscriber for hooks
+
+<br/>
+
+## [0.1.4-beta] - 2024-03-15 _00:00_
 
 - Fix imports